Conflicts:
	distrib/win32/setup.iss.template
	doomsday/CMakeLists.txt
	doomsday/build/win32/doomsday_cl.rsp
	doomsday/engine/api/dd_share.h
	doomsday/engine/api/dd_vectorgraphic.h
	doomsday/engine/api/doomsday.def
	doomsday/engine/api/doomsday.h
	doomsday/engine/portable/include/b_command.h
	doomsday/engine/portable/include/b_context.h
	doomsday/engine/portable/include/b_device.h
	doomsday/engine/portable/include/b_main.h
	doomsday/engine/portable/include/b_util.h
	doomsday/engine/portable/include/con_busy.h
	doomsday/engine/portable/include/con_decl.h
	doomsday/engine/portable/include/dd_version.h
	doomsday/engine/portable/include/de_base.h
	doomsday/engine/portable/include/de_graphics.h
	doomsday/engine/portable/include/de_infine.h
	doomsday/engine/portable/include/fi_main.h
	doomsday/engine/portable/include/gl_defer.h
	doomsday/engine/portable/include/gl_font.h
	doomsday/engine/portable/include/gl_hq2x.h
	doomsday/engine/portable/include/gl_png.h
	doomsday/engine/portable/include/m_filehash.h
	doomsday/engine/portable/include/m_profiler.h
	doomsday/engine/portable/include/m_string.h
	doomsday/engine/portable/include/net_main.h
	doomsday/engine/portable/include/p_material.h
	doomsday/engine/portable/include/r_extres.h
	doomsday/engine/portable/include/r_main.h
	doomsday/engine/portable/include/sv_infine.h
	doomsday/engine/portable/include/sys_system.h
	doomsday/engine/portable/include/ui2_main.h
	doomsday/engine/portable/include/ui_panel.h
	doomsday/engine/portable/src/b_context.c
	doomsday/engine/portable/src/b_device.c
	doomsday/engine/portable/src/b_main.c
	doomsday/engine/portable/src/b_util.c
	doomsday/engine/portable/src/bsp_map.c
	doomsday/engine/portable/src/cl_main.c
	doomsday/engine/portable/src/cl_mobj.c
	doomsday/engine/portable/src/cl_oldworld.c
	doomsday/engine/portable/src/cl_player.c
	doomsday/engine/portable/src/cl_world.c
	doomsday/engine/portable/src/con_main.c
	doomsday/engine/portable/src/dam_main.c
	doomsday/engine/portable/src/dd_loop.c
	doomsday/engine/portable/src/dd_main.c
	doomsday/engine/portable/src/def_read.c
	doomsday/engine/portable/src/dgl_texture.c
	doomsday/engine/portable/src/gl_defer.c
	doomsday/engine/portable/src/gl_font.c
	doomsday/engine/portable/src/gl_main.c
	doomsday/engine/portable/src/gl_png.c
	doomsday/engine/portable/src/gl_texmanager.c
	doomsday/engine/portable/src/m_filehash.c
	doomsday/engine/portable/src/m_misc.c
	doomsday/engine/portable/src/m_string.c
	doomsday/engine/portable/src/net_buf.c
	doomsday/engine/portable/src/net_main.c
	doomsday/engine/portable/src/p_bmap.c
	doomsday/engine/portable/src/p_control.c
	doomsday/engine/portable/src/p_dmu.c
	doomsday/engine/portable/src/p_material.c
	doomsday/engine/portable/src/p_materialmanager.c
	doomsday/engine/portable/src/p_mobj.c
	doomsday/engine/portable/src/p_sight.c
	doomsday/engine/portable/src/p_ticker.c
	doomsday/engine/portable/src/r_data.c
	doomsday/engine/portable/src/r_extres.c
	doomsday/engine/portable/src/r_lumobjs.c
	doomsday/engine/portable/src/r_main.c
	doomsday/engine/portable/src/r_things.c
	doomsday/engine/portable/src/r_world.c
	doomsday/engine/portable/src/rend_bias.c
	doomsday/engine/portable/src/rend_decor.c
	doomsday/engine/portable/src/rend_fakeradio.c
	doomsday/engine/portable/src/rend_sprite.c
	doomsday/engine/portable/src/s_environ.c
	doomsday/engine/portable/src/sv_main.c
	doomsday/engine/portable/src/sv_pool.c
	doomsday/engine/portable/src/sys_direc.c
	doomsday/engine/portable/src/sys_filein.c
	doomsday/engine/portable/src/sys_sdl_window.c
	doomsday/engine/portable/src/ui_mpi.c
	doomsday/engine/unix/include/sys_findfile.h
	doomsday/engine/unix/src/dd_uinit.c
	doomsday/engine/unix/src/sys_dylib.c
	doomsday/engine/win32/include/dd_winit.h
	doomsday/engine/win32/src/dd_winit.c
	doomsday/engine/win32/src/sys_console.c
	doomsday/engine/win32/src/sys_findfile.c
	doomsday/engine/win32/src/sys_input.c
	doomsday/engine/win32/src/sys_window.c
	doomsday/plugins/common/include/d_netsv.h
	doomsday/plugins/common/include/f_infine.h
	doomsday/plugins/common/include/hu_lib.h
	doomsday/plugins/common/include/hu_log.h
	doomsday/plugins/common/include/hu_menu.h
	doomsday/plugins/common/include/hu_msg.h
	doomsday/plugins/common/include/hu_stuff.h
	doomsday/plugins/common/include/m_defs.h
	doomsday/plugins/common/include/p_automap.h
	doomsday/plugins/common/include/st_lib.h
	doomsday/plugins/common/src/am_map.c
	doomsday/plugins/common/src/d_net.c
	doomsday/plugins/common/src/d_netcl.c
	doomsday/plugins/common/src/f_infine.c
	doomsday/plugins/common/src/g_eventsequence.c
	doomsday/plugins/common/src/g_game.c
	doomsday/plugins/common/src/hu_lib.c
	doomsday/plugins/common/src/hu_log.c
	doomsday/plugins/common/src/hu_menu.c
	doomsday/plugins/common/src/hu_pspr.c
	doomsday/plugins/common/src/hu_stuff.c
	doomsday/plugins/common/src/m_multi.c
	doomsday/plugins/common/src/p_automap.c
	doomsday/plugins/common/src/p_mapsetup.c
	doomsday/plugins/common/src/p_mapspec.c
	doomsday/plugins/common/src/p_start.c
	doomsday/plugins/common/src/p_user.c
	doomsday/plugins/common/src/rend_automap.c
	doomsday/plugins/common/src/st_lib.c
	doomsday/plugins/dehread/src/dehmain.c
	doomsday/plugins/directsound/src/driver_directsound.cpp
	doomsday/plugins/exampleplugin/src/example.c
	doomsday/plugins/jdoom/include/d_console.h
	doomsday/plugins/jdoom/include/mn_def.h
	doomsday/plugins/jdoom/include/version.h
	doomsday/plugins/jdoom/include/wi_stuff.h
	doomsday/plugins/jdoom/src/d_api.c
	doomsday/plugins/jdoom/src/d_main.c
	doomsday/plugins/jdoom/src/d_refresh.c
	doomsday/plugins/jdoom/src/p_mobj.c
	doomsday/plugins/jdoom/src/p_pspr.c
	doomsday/plugins/jdoom/src/st_stuff.c
	doomsday/plugins/jdoom/src/wi_stuff.c
	doomsday/plugins/jdoom64/include/d_api.h
	doomsday/plugins/jdoom64/include/dstrings.h
	doomsday/plugins/jdoom64/include/mn_def.h
	doomsday/plugins/jdoom64/include/version.h
	doomsday/plugins/jdoom64/include/wi_stuff.h
	doomsday/plugins/jdoom64/src/d_api.c
	doomsday/plugins/jdoom64/src/d_main.c
	doomsday/plugins/jdoom64/src/p_mobj.c
	doomsday/plugins/jdoom64/src/st_stuff.c
	doomsday/plugins/jdoom64/src/wi_stuff.c
	doomsday/plugins/jheretic/include/doomdef.h
	doomsday/plugins/jheretic/include/mn_def.h
	doomsday/plugins/jheretic/include/version.h
	doomsday/plugins/jheretic/src/h_api.c
	doomsday/plugins/jheretic/src/h_main.c
	doomsday/plugins/jheretic/src/p_mobj.c
	doomsday/plugins/jheretic/src/st_stuff.c
	doomsday/plugins/jhexen/include/h2def.h
	doomsday/plugins/jhexen/include/mn_def.h
	doomsday/plugins/jhexen/include/version.h
	doomsday/plugins/jhexen/include/x_api.h
	doomsday/plugins/jhexen/include/x_main.h
	doomsday/plugins/jhexen/src/h2_main.c
	doomsday/plugins/jhexen/src/hrefresh.c
	doomsday/plugins/jhexen/src/p_mobj.c
	doomsday/plugins/jhexen/src/st_stuff.c
	doomsday/plugins/jhexen/src/x_api.c
	doomsday/plugins/winmm/src/cdaudio.cpp
	doomsday/plugins/winmm/src/dswinmm.cpp

+++ doomsday/engine/portable/src/b_context.c
@@ -230,7 +230,7 @@ void B_UpdateDeviceStateAssociations(void)
         inputdev_t* dev = I_GetDevice(i, false);
 
         // Keys.
-        for(j = 0; j < dev->numKeys; ++j)
+        for(j = 0; j < (int)dev->numKeys; ++j)
         {
             if(dev->keys[j].assoc.bContext != dev->keys[j].assoc.prevBContext &&
                dev->keys[j].isDown)
@@ -241,7 +241,7 @@ void B_UpdateDeviceStateAssociations(void)
         }
 
         // Axes.
-        for(j = 0; j < dev->numAxes; ++j)
+        for(j = 0; j < (int)dev->numAxes; ++j)
         {
             if(dev->axes[j].assoc.bContext != dev->axes[j].assoc.prevBContext &&
                dev->axes[j].position != 0)
@@ -252,7 +252,7 @@ void B_UpdateDeviceStateAssociations(void)
         }
 
         // Hats.
-        for(j = 0; j < dev->numHats; ++j)
+        for(j = 0; j < (int)dev->numHats; ++j)
         {
             if(dev->hats[j].assoc.bContext != dev->hats[j].assoc.prevBContext &&
                dev->hats[j].pos >= 0)
